Reorganized and make components respond to change in actions
