+++ packages/tools/vsm/src/components/tools/useSceneNode.ts
@@ -5,7 +5,7 @@
 import type { AbstractMesh } from "core/Meshes/abstractMesh";
 import type { Nullable } from "core/types";
 import { useContext } from "react";
-import { SceneContext } from "../../SceneContext";
+import { SceneContext } from "../../context/SceneContext";
 
 export function useSceneNode(nodeName: string): Nullable<AbstractMesh> | undefined {
     const { scene } = useContext(SceneContext);
