+++ packages/tools/vsm/src/components/tools/useSelectedAction.ts
@@ -1,7 +1,13 @@
+import type { Nullable } from "core/types";
 import { useContext } from "react";
-import { ActionSelectionContext } from "../ActionSelectionContext";
+import type { BaseAction } from "../../actions/actions/BaseAction";
+import { ActionSelectionContext } from "../../context/ActionSelectionContext";
 
 export const useSelectedAction = () => {
-    const { selectedAction, setSelectedAction } = useContext(ActionSelectionContext);
-    return { selectedAction, setSelectedAction };
+    const { selectedActionWrapper, setSelectedActionWrapper } = useContext(ActionSelectionContext);
+    const { action, lastUpdate } = selectedActionWrapper;
+    const setSelectedAction = (newAction: Nullable<BaseAction>) => {
+        setSelectedActionWrapper({ action: newAction, lastUpdate: Date.now() });
+    };
+    return { selectedAction: action, lastUpdate, setSelectedAction };
 };
