+++ packages/tools/vsm/src/components/tools/useSelectedState.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { StateSelectionContext } from "../StateSelectionContext";
+import { StateSelectionContext } from "../../context/StateSelectionContext";
 
 export const useSelectedState = () => {
     const { selectedState, setSelectedState } = useContext(StateSelectionContext);
