+++ packages/tools/vsm/src/context/SceneContext.ts
